+++ R/import_sddf_country.R
@@ -2,15 +2,17 @@
 #'
 #' @param country a character of length 1 with the full name of the country. 
 #' Use \code{\link{show_countries}} for a list of available countries.
-#' @param rounds a numeric vector with the rounds to download. See \code{\link{show_rounds}}
-#' for all available rounds.
+#' 
+#' @param rounds a numeric vector with the rounds to download. See \code{\link{show_sddf_rounds}}
+#' for all available rounds for any given country.
+#' 
 #' @param ess_email a character vector with your email, such as "your_email@email.com".
 #' If you haven't registered in the ESS website, create an account at 
 #' \url{http://www.europeansocialsurvey.org/user/new}. A preferred method is to login
 #' through \code{\link{set_email}}.
 #' 
 #' @param output_dir a character vector with the output directory in case you want to
-#' only download the files using \code{download_country}. Defaults to your working
+#' only download the files using \code{download_sddf_country}. Defaults to your working
 #' directory. This will be interpreted as a \strong{directory} and not a path with
 #' a file name.
 #'
@@ -19,8 +21,8 @@
 #' @details
 #'
 #' SDDF data (Sample Design Data Files) are data sets that contain additional columns with the
-#' sample design and weights for a given country in a given rounds. These additional columns are
-#' required to perform any weighted analysis of the ESS data. Users interested in using this data
+#' sample design and weights for a given country in a given round. These additional columns are
+#' required to perform any complex weighted analysis of the ESS data. Users interested in using this data
 #' should read the description of SDDF files \href{http://www.europeansocialsurvey.org/methodology/ess_methodology/sampling.html}{here}
 #' and should read \href{http://www.europeansocialsurvey.org/data/download_sample_data.html}{here} for the
 #' sampling design of the country of analysis for that specific round.
@@ -36,8 +38,8 @@
 #' so that the user can switch between formats if any encoding errors are found in the data. For more
 #' details see the discussion \href{https://github.com/ropensci/essurvey/issues/11}{here}.
 #' 
-#' Given that the SDDF data is not very complete, some countries do not have SDDF data
-#' in Stata or SPSS. For that reason, the \code{format} argument is not used in \code{import_sddf_country}.
+#' Additionally, given that the SDDF data is not very complete, some countries do not have SDDF data
+#' in Stata or SPSS formats. For that reason, the \code{format} argument is not used in \code{import_sddf_country}.
 #' Internally, \code{Stata} is chosen over \code{SPSS} and \code{SPSS} over \code{SAS} in that
 #' order of preference.
 #' 
@@ -49,13 +51,12 @@
 #' Starting from round 7 (including), the ESS switched the layout of SDDF data.
 #' Before the rounds, SDDF data was published separately by wave-country
 #' combination. From round 7 onwards, all SDDF data is released as a single
-#' integrated file with all countries combined. \code{import_sddf_country}
+#' integrated file with all countries combined for that given round. \code{import_sddf_country}
 #' takes care of this nuance by reading the data and filtering the chosen
-#' country. \code{download_sddf_country} downloads the raw file but also
-#' reads in memory to subset the specific country requested. This shouldn't
-#' raise any problems but beware that reading/writing the data might delete
-#' some properties such as wrongly saved labels. For this reason, 'sas' format
-#' is not supported for this specific function.
+#' country automatically. \code{download_sddf_country} downloads the raw file but also
+#' reads the data into memory to subset the specific country requested. This
+#' process should be transparent to the user but beware that reading/writing the data might delete
+#' some of it's properties such as dropping the labels or label attribute.
 #'
 #' @return for \code{import_sddf_country} if \code{length(rounds)} is 1, it returns a tibble with
 #' the latest version of that round. Otherwise it returns a list of \code{length(rounds)}
@@ -69,10 +70,8 @@
 #' 
 #' set_email("your_email@email.com")
 #' 
-#' # Get first three rounds for Denmark
 #' sp_three <- import_sddf_country("Spain", 5:6)
 #' 
-#' # For now, only SDDF files from rounds 1:6 are implemented in the package
 #' show_sddf_rounds("Spain")
 #' 
 #' # Only download the files, this will return nothing
@@ -88,16 +87,19 @@
 #' # By default, download_sddf_country downloads 'stata' files but
 #' # you can also download 'spss' or 'sas' files.
 #' 
-#' download_country(
+#' download_sddf_country(
 #'  "Spain",
-#'  rounds = 5:6,
+#'  rounds = 1:8,
 #'  output_dir = temp_dir,
 #'  format = 'spss'
 #' )
 #' 
 #' }
 #' 
-import_sddf_country <- function(country, rounds, ess_email = NULL, format = NULL) {
+import_sddf_country <- function(country,
+                                rounds,
+                                ess_email = NULL,
+                                format = NULL) {
 
   stopifnot(is.character(country), length(country) > 0)
   stopifnot(is.numeric(rounds), length(rounds) > 0)
@@ -180,71 +182,87 @@ download_sddf_country <- function(country,
       "You cannot read SAS but only 'spss' and 'stata' files with this function. See ?download_sddf_country for more details") # nolint
   }
 
-
+  rounds <- sort(rounds)
   late_rounds <- rounds > 6
   
-  if (any(late_rounds)) {
+  if (all(late_rounds)) {
 
-    urls <- c(country_url_sddf(country, rounds[!late_rounds], format),
-              country_url_sddf_late_rounds(country, rounds[late_rounds], format)
-              )
-  } else {
+    urls <- country_url_sddf_late_rounds(country, rounds[late_rounds], format)
+    
+
+  } else if (all(!late_rounds)) {
 
     urls <- country_url_sddf(country, rounds, format)
 
+  } else {
+    urls <- c(country_url_sddf(country, rounds[!late_rounds], format),
+              country_url_sddf_late_rounds(country, rounds[late_rounds], format)
+              )
   }
 
+  dir_download <-
+    download_format(urls = urls,
+                    country = country,
+                    ess_email = ess_email,
+                    only_download = TRUE,
+                    output_dir = output_dir)
+
+  # If there are any late rounds, we need to read them back to memory
+  # subset the specific country they requested and resave them back
+  # to the same directory they were read from
+  if (any(late_rounds)) {
+
+    all_data <-
+      suppress_all(
+        read_format_data(dir_download[late_rounds], rounds[late_rounds])
+      )
+
+    # Turn to a list in case it's only one round
+    all_data <- if (!inherits(all_data, "list")) list(all_data) else all_data
+
+    # Search for the 2 letter code because we need to subset
+    # from the integrated SDDF for the current country
+    country_code <- country_lookup[country]
+
+    # Subset the selected country from the integrated late rounds
+    all_data <- lapply(all_data, function(x) x[x$cntry == country_code, ])
+
+    format_ext <- c(".dta", ".sav", ".por")
+    # Get all paths from the format
+    # This **should be** the same paths used to read the data because
+    # these exact same lines of code are used in read_format_data.
+    format_dirs <- list.files(dir_download[late_rounds],
+                              pattern = paste0(format_ext, "$", collapse = "|"),
+                              full.names = TRUE)
+
+    if (length(format_dirs) != length(all_data)) {
+      # If this bug happens it means that there are more files with
+      # stata/spss/sas extensions than data already read. This means
+      # that could save the data in an otherwise no valid file
+      # messing up the logic of which file hosts the subsetted data.
+      stop("This is an internal check for a possible bug with the ESS data.
+            Please file an issue at https://github.com/ropensci/essurvey/issues
+            with exact line of code that generated the error. For example:
+            download_country_sddf('Italy', 1:3)")
+    }
+    
+    # Save only the .dta/.sav/.por files
+    for (.x in seq_along(format_dirs)) {
+      
+      # Use function to read the specified format
+      format_save <-
+        switch(file_ext(format_dirs[[.x]]),
+               "dta" = haven::write_dta,
+               "por" = haven::write_sav,
+               "sav" = haven::write_sav
+               )
+
+      # .por files give a lot of trouble. Here we always read them with
+      # read/write_sav so we change the file extension to .sav to save them
+      # as .sav
+      format_save(all_data[[.x]], gsub(".por", ".sav", format_dirs[.x]))
+    }
+  }
   
-  invisible(
-    dir_download <-
-      download_format(urls = urls,
-                      country = country,
-                      ess_email = ess_email,
-                      only_download = TRUE,
-                      output_dir = output_dir)
-  )
-
-  ## if (any(late_rounds)) {
-
-  ##   downloaded_rounds <-
-  ##     as.numeric(
-  ##       gsub(
-  ##         "ESS",
-  ##         "",
-  ##         stringr::str_extract(dir_download, "ESS[0-9]{1,}")
-  ##       )
-  ##     )
-
-  ##   new_late_rounds <- downloaded_rounds > 6
-
-  ##   all_data <-
-  ##     suppress_all(
-  ##       read_format_data(dir_download[new_late_rounds], rounds[late_rounds])
-  ##     )
-
-  ## # Search for the 2 letter code because we need to subset
-  ## # from the integrated SDDF for the current country
-  ## country_code <- country_lookup[country]
-
-  ## # Subset the selected country from the integrated late rounds
-  ## all_data <- lapply(all_data, function(x) x[x$cntry == country_code, ])
-
-  ## format_ext <- c(".dta", ".sav", ".por")
-  ## # Get all paths from the format
-  ## format_dirs <- list.files(dir_download[new_late_rounds],
-  ##                           pattern = paste0(format_ext, "$", collapse = "|"),
-  ##                           full.names = TRUE)
-  
-  ## # Read only the .dta/.sav/.por files
-  ## dataset <- lapply(format_dirs, function(.x) {
-  
-  ##   # Use function to read the specified format
-  ##   format_read <-
-  ##     switch(file_ext(.x),
-  ##            "dta" = haven::read_dta,
-  ##            "por" = haven::read_sav,
-  ##            "sav" = haven::read_sav
-  ##            )
-
-
+  invisible(dir_download)
 }